+++ examples/render/serverside.js
@@ -5,7 +5,12 @@ const view = require(path.join(__dirname, 'sample-view.json'))
 
 const canvas = new Canvas()
 
-canvas.render(view, function (err, outputSVG) {
+canvas.load(view)
+
+const width = 600
+const height = 400
+
+canvas.toSVG({ width, height }, (err, outputSVG) => {
   if (err) throw err
 
   console.log(outputSVG)